Created DevKit V1 build target, use TX16S simulator widget.
LED_RED is connected on DevKit to PA5 instead of PE2 as in TX16S. PE2 is ETM trace pin.

PCBREV1 and PCBREV2 are usable on DevKit only when ETM is not used, as the pins are shared with trace (PE4 & PE5).

Switch E front/low mapped on DevKit to PH7 instead of PE3, which is needed for TRACEDATA0 (and alternatively Switch J if ETM is not used) on DevKit.

Following pins are NOT usable together with ETM:
* PE3 as Switch J
* PE6 as Haptic
Both are required for trace.

On DevKit Blue LED uses PB2 instead of PE5 (latter is used for ETM on DevKit).

DevKit uses PH8 for green LED instead of PE4.

DevKit v1 uses PH13 as USART3_PWR instead of PA15 that can be used for TDI on DevKit.

Use PH15 for external module power, freeing PB3 for optional SWO (simple trace output).